Move components into folder
